Handle copy_if_larger edge case

+++ core/worker.py
@@ -444,7 +444,11 @@ class Worker(QRunnable):
                     os.rename(self.output, self.final_output)
 
                 # Copy original
-                if self.settings["copy_if_larger"] and os.path.getsize(self.org_item_abs_path) < os.path.getsize(self.final_output):
+                if (
+                    self.settings["copy_if_larger"] and
+                    os.path.getsize(self.org_item_abs_path) < os.path.getsize(self.final_output) and
+                    self.params["format"] not in ("Lossless JPEG Recompression", "JPEG Reconstruction")
+                ):
                     os.remove(self.final_output)
                     self.final_output = getUniqueFilePath(self.output_dir, self.item_name, self.item_ext, False)
                     shutil.copy(self.org_item_abs_path, self.final_output)
